sua lai cau truc code thanh OOP code
